Initial stab at engine
The trick is to view the interface and s tate as two separate (not really) worlds. In the sense that, they will both be attuned to each other - but the power will lie in communication over a high-level bus where interface events translate to higher-level state updates, which in turn translate to interface changes.

While writing, I think it's worth it to introduce some kind of `Window`-abstraction.